% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CouFinder.R
\name{CouFinder}
\alias{CouFinder}
\title{A Country finder.}
\usage{
CouFinder(Region)
}
\arguments{
\item{region}{Write down any region you would like! (as long as it's in the database that is)}
}
\value{
The countries that are part of your region.
}
\description{
This function allows you to find out what region your country of choice resides in.
}
\examples{
CouFinder("x")
}
