% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Life_expectancy_graph.R
\name{Life_expectancy_graph}
\alias{Life_expectancy_graph}
\title{A graph that shows the life expectancy in a specific year, in a specific region.}
\usage{
Life_expectancy_graph(region_wanted, year_wanted)
}
\arguments{
\item{continent, year}{Make sure to write down both the year AND region wanted. The year can only be between 2002 and 2015.}
}
\value{
A bar graph showing the difference in life expectancy between the countries.
}
\description{
This function allows you to find out what the difference is between the life expectancy, taken from countries all residing in the same region, in the same year.
}
\examples{
life_expectancy_graph("x","y")
}
