% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionFinder.R
\name{RegionFinder}
\alias{RegionFinder}
\title{A region finder.}
\usage{
RegionFinder(country)
}
\arguments{
\item{country}{Write down any country you would like! (as long as it's in the database that is)}
}
\value{
The region your country resides in.
}
\description{
This function allows you to find out what region your country of choice resides in.
}
\examples{
RegionFinder("x")
}
